fix(angular-rspack): fix i18n inlining
