+++ e2e/fixtures/rspack-csr-i18n/src/main.ts
@@ -1,5 +1,6 @@
 /// <reference types="@angular/localize" />
 import '@angular/common/locales/global/fr';
+import '@angular/common/locales/global/es';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { appConfig } from './app/app.config';
 import { AppComponent } from './app/app.component';
